+++ mycli/types.py
@@ -1,4 +1,11 @@
 from collections import namedtuple
+from typing import Literal
 
 # Query tuples are used for maintaining history
 Query = namedtuple("Query", ["query", "successful", "mutating"])
+
+OutputMode = Literal[
+    'explorer',
+    'expanded',
+    'tabular',
+]